+++ examples/resources/osc_parameter/import.sh
@@ -1,0 +1,2 @@
+# <parameter_store>/<key>
+terraform import osc_parameter.node_env myapiconfig/NODE_ENV